upd `garment.py` - users can only delete their own garments

+++ app/routers/garment.py
@@ -49,6 +49,10 @@ async def update_garment(id: int, garment_in: schemas.GarmentBase, db: db_depend
 async def delete_garment(id: int, db: db_dependency, current_user: models.User = Depends(oauth2.get_current_user)):
     try:
         garment = db.query(models.Garment).filter(models.Garment.id == id).one()
+
+        if garment.user_id != current_user.id:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+
         db.delete(garment)
         db.commit()
         return {"message": f"Garment {id} deleted successfully"}
